apps/browser/src/autofill/overlay/abstractions/autofill-overlay-iframe.service.ts
type AutofillOverlayIframeExtensionMessage = {
  command: string;
  styles?: Partial<CSSStyleDeclaration>;
  theme?: string;
};

type AutofillOverlayIframeWindowMessageHandlers = {
  [key: string]: CallableFunction;
  updateAutofillOverlayListHeight: (message: AutofillOverlayIframeExtensionMessage) => void;
};

type AutofillOverlayIframeExtensionMessageParam = {
  message: AutofillOverlayIframeExtensionMessage;
};

type BackgroundPortMessageHandlers = {
  [key: string]: CallableFunction;
  initAutofillOverlayList: ({ message }: AutofillOverlayIframeExtensionMessageParam) => void;
  updateIframePosition: ({ message }: AutofillOverlayIframeExtensionMessageParam) => void;
  updateOverlayHidden: ({ message }: AutofillOverlayIframeExtensionMessageParam) => void;
};

interface AutofillOverlayIframeService {
  initOverlayIframe(initStyles: Partial<CSSStyleDeclaration>, ariaAlert?: string): void;
}

export {
  AutofillOverlayIframeExtensionMessage,
  AutofillOverlayIframeWindowMessageHandlers,
  BackgroundPortMessageHandlers,
  AutofillOverlayIframeService,
};
